refactor(analysis-types): move complexity serde tests
